pyproject.toml
[project]
name = "python-api-elevenlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.34.0",
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
]
